Separate mysql storable implementation into dedicated module
